bind holder

+++ app/src/main/java/com/example/recyclerview08012024/Cinema.kt
@@ -5,5 +5,5 @@ class Cinema (
     var name: String,
     var address: String,
     var phone: String,
-    var distance: Float
+    var distance: Int
 )
